This version of Antimony is now broken--I'm only committing it so I can more easily work on it from home.  The intent is to convert the return functions from using local cached values to using malloc and strdup--this isn't yet done.

+++ simpleantimonyapi.h
@@ -5,8 +5,8 @@
 
 BEGIN_C_DECLS;
 
-LIB_EXTERN int getNumFloatingSpecies (char* moduleName);
-LIB_EXTERN int getNumBoundarySpecies (char* moduleName);
+LIB_EXTERN size_t getNumFloatingSpecies (char* moduleName);
+LIB_EXTERN size_t getNumBoundarySpecies (char* moduleName);
 
 LIB_EXTERN const char* getNthFloatingSpeciesName (char* moduleName, size_t n); 
 LIB_EXTERN const char* getNthBoundarySpeciesName (char* moduleName, size_t n); 
@@ -15,16 +15,16 @@ LIB_EXTERN const char** getFloatingSpeciesList (const char* moduleName);
 LIB_EXTERN const char** getBoundarySpeciesList (const char* moduleName);
 
 LIB_EXTERN const char** getReactionNameList (const char* moduleName);
-LIB_EXTERN int getNumKineticLaws (char* moduleName);
+LIB_EXTERN size_t getNumKineticLaws (char* moduleName);
 LIB_EXTERN const char* getNthKineticLaw (char* moduleName, size_t n);
 
-LIB_EXTERN int getNumberReactants (char* moduleName, int reactionId);
-LIB_EXTERN int getNumberProducts (char* moduleName, int reactionId);
+LIB_EXTERN size_t getNumberReactants (char* moduleName, size_t reactionId);
+LIB_EXTERN size_t getNumberProducts (char* moduleName, size_t reactionId);
 
-LIB_EXTERN int getNthReactantStoichiometry (char* moduleName, int reactionId, int reactantId);
-LIB_EXTERN int getNthProductStoichiometry (char* moduleName, int reactionId, int productId);
+LIB_EXTERN size_t getNthReactantStoichiometry (char* moduleName, size_t reactionId, size_t reactantId);
+LIB_EXTERN size_t getNthProductStoichiometry (char* moduleName, size_t reactionId, size_t productId);
 
-LIB_EXTERN void freeMatrix (double** mat, int nRows);
+LIB_EXTERN void freeMatrix (double** mat, size_t nRows);
 LIB_EXTERN double** myGetStoichiometryMatrix (char* moduleName);
 
 END_C_DECLS;
